# swagger/swagger.yml
openapi: 3.1.0
info:
  title: Servicio de Datos - OpenAPI 3.1
  description: |-
    Servicio de datos para el sistema software agro-predict.   
    Encargado de solicitar, acumular, normalizar y almacenar en la base de datos, datos obtenidos por servicios públicos externos.  

    Devuelve los datos en formato JSON con codificación UTF-8.

    Links sobre los servicios de abastecimiento de datos meteorológicos:
    - [Servicio AEMET](https://github.com/agro-predict-tfg-2026/Aemet.git)
    - [Servicio SiAR](https://github.com/agro-predict-tfg-2026/SiAR.git)
    - [Servicio ITACyL](https://github.com/agro-predict-tfg-2026/ITACyL.git)
    - [Servicio IA Generativa](https://github.com/agro-predict-tfg-2026/mcp-ia-service.git)

  contact:
    email: amendoma@alumnos.unex.es
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:5000
  - url: http://data-service:9002
  - url: http://51.44.230.198:9002
...
paths:
...
  /crop/variedades/{nombre}/umbrales:
    get:
      summary: Obtener umbrales de una variedad
      description: Devuelve los umbrales de riesgo de helada definidos para una variedad concreta.
      tags:
        - Cultivos
      parameters:
        - name: nombre
          in: path
          required: true
          description: Nombre de la variedad
          schema:
            type: string
            example: "Guara"
      responses:
        "200":
          description: Umbrales de la variedad obtenidos correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UmbralesRiesgo'
        "404":
          description: No se han encontrado umbrales para la variedad indicada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: "Los umbrales para la variedad X no están definidos"
                status: 404
                error: "Data Not Found"
        "502":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: "Provider Error"
                status: 502
                error: "Error del proveedor de datos"

    post:
      summary: Registrar umbrales para una variedad
      description: Crea los umbrales de riesgo de helada para una variedad concreta.
      tags:
        - Cultivos
      parameters:
        - name: nombre
          in: path
          required: true
          description: Nombre de la variedad
          schema:
            type: string
            example: "Guara"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UmbralesRiesgo'
      responses:
        "201":
          description: Umbral creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessCreated'
        "400":
          description: Datos en JSON no proporcionados o inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
              example:
                message: "Datos en json requeridos"
                status: 400
                error: "Bad Request"
        "502":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
              example:
                message: "Provider Error"
                status: 502
                error: "Error del proveedor de datos"
  /crop/variedades/{nombre}/horas_frio:
    get:
      summary: Obtener horas de frío de una variedad
      description: Devuelve los umbrales de horas de frío (mínimo y máximo) definidos para una variedad.
      tags:
        - Cultivos
      parameters:
        - name: nombre
          in: path
          required: true
          description: Nombre de la variedad
          schema:
            type: string
            example: "Guara"
      responses:
        "200":
          description: Horas de frío de la variedad obtenidas correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VariedadHorasFrio'
        "404":
          description: No se han encontrado datos de horas de frío para la variedad indicada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: "No se han encontrado datos hora_frio para la variedad X"
                status: 404
                error: "Data Not Found"
        "502":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: "Provider Error"
                status: 502
                error: "Error del proveedor de datos"

    patch:
      summary: Actualizar horas de frío de una variedad
      description: Actualiza los valores de horas de frío mínimas y/o máximas de una variedad.
      tags:
        - Cultivos
      parameters:
        - name: nombre
          in: path
          required: true
          description: Nombre de la variedad
          schema:
            type: string
            example: "Guara"
        - name: frio_max
          in: query
          required: false
          description: Nuevo valor máximo de horas de frío
          schema:
            type: integer
            example: 650
        - name: frio_min
          in: query
          required: false
          description: Nuevo valor mínimo de horas de frío
          schema:
            type: integer
            example: 420
      responses:
        "200":
          description: Horas de frío actualizadas correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessOk'
        "400":
          description: No se han proporcionado los nuevos valores a actualizar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: No se puede actualizar las horas frio de la variedad X porque no se han definido sus nuevos valores (frio_max, frio_min)
                status: 400
                error: "Invalid Parameters"
        "502":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: "Provider Error"
                status: 502
                error: "Error del proveedor de datos"
  /crop/modelos/{codigo}/variedades:
    get:
      summary: Obtener variedades por modelo de horas de frío
      description: Devuelve las variedades de cultivo que implementan el modelo de horas de frío indicado.
      tags:
        - Cultivos
      parameters:
        - name: codigo
          in: path
          required: true
          description: Código del modelo de horas de frío
          schema:
            type: string
            enum: ['CM7', 'C07', 'UF', 'PF']
      responses:
        "200":
          description: Variedades del modelo obtenidas correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VariedadesModelos'
        "400":
          description: Se ha producido un error en la validación de parámetros
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: "Parametros de entrada incorrectos"
                status: 400
                error: "Invalid Parameters"
        "404":
          description: No se han encontrado variedades para el modelo indicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: "No se han encontrado variedades que implementen el modelo X"
                status: 404
                error: "Data Not Found"
        "502":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              exampele:
                message: "Provider Error"
                status: 502
                error: "Error del proveedor de datos"
  /crop/plague:
    get:
      summary: Obtener cultivos con sus plagas asociadas
      description: >
        Devuelve los cultivos indicados junto con sus plagas y enfermedades asociadas.
        Se pueden indicar múltiples cultivos separados por comas.
      tags:
        - Cultivos
      parameters:
        - name: cultivos
          in: query
          required: true
          description: Nombre o nombres de cultivo separados por comas
          schema:
            type: string
            example: "Maiz"
      responses:
        "200":
          description: Cultivos con plagas asociadas obtenidos correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrediccionPlagasCalculadas'
        "400":
          description: Parámetro cultivos no indicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: "Se debe indicar el nombre de el/los cultivos a analizar"
                status: 400
                error: "Invalid Parameters"
        "404":
          description: Datos de plagas asociadas a cultivos no encuntrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: "No se han encontrado datos de plagas asociados al cultivo indicado"
                status: 404
                error: "Data Not Found"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:  
                message: "Provider Error"
                status: 500
                error: "Error del proveedor de datos"
  /crop/parcel:
    get:
      tags:
        - Cultivos
      summary: Obtener asociación cultivo-parcela
      parameters:
        - name: parcela
          in: query
          required: false
          schema:
            type: string

        - name: cultivo
          in: query
          required: false
          schema:
            type: string

      responses:
        "200":
          description: Cultivos con parcela asociada obtenidos correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrediccionPlagasCalculadas'
        "400":
          description: Parámetro cultivos o parcela no indicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: "Se debe indicar el nombre de el/los cultivos o la parcela a analizar"
                status: 400
                error: "Invalid Parameters"
        "404":
          description: Datos de parcelas asociadas a cultivos no encuntrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: "No se han encontrado datos de plagas asociados al cultivo indicado"
                status: 404
                error: "Data Not Found"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:  
                message: "Provider Error"
                status: 500
                error: "Error del proveedor de datos"
  /crop/{campo}:
    get:
      summary: Listar recursos de cultivo
      description: >
        Endpoint genérico para obtener listados de recursos relacionados con cultivos.
        Los valores válidos de `campo` son: `variedades`, `modelos`, `etapas_fenologicas`, `cultivos`.
        El recurso `variedades` admite el parámetro opcional `cultivo` para filtrar por cultivo.
      tags:
        - Cultivos
      parameters:
        - name: campo
          in: path
          required: true
          description: Tipo de recurso a consultar
          schema:
            type: string
            enum: [variedades, modelos, etapas_fenologicas, cultivos]
            example: "variedades"
        - name: cultivo
          in: query
          required: false
          description: Filtro por nombre de cultivo (solo aplica cuando campo = variedades)
          schema:
            type: string
            example: "Almendro"
      responses:
        "200":
          description: Listado del recurso solicitado obtenido correctamente
          content:
            application/json:
              schema:
                type: array
                description: Estructura variable según el campo solicitado
                items:
                  $ref: '#/components/schemas/VariedadesModelos'
        "400":
          description: Campo no soportado o parámetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: "Parametros invalidos, se debe especificar el campo"
                status: 400
                error: "Invalid Parameters"
        "404":
          description: No se han encontrado datos para el recurso solicitado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: "No se encontraron datos de X"
                status: 404
                error: "Data Not Found"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example: 
                message: "Provider Error"
                status: 500
                error: "Error del proveedor de datos"

    post:
      summary: Crear un nuevo recurso de cultivo
      description: >
        Endpoint genérico para crear nuevos recursos relacionados con cultivos.
        Los valores válidos de `campo` son: `variedades`, `modelos`, `cultivos`.
        El recurso `etapas_fenologicas` no admite creación.
      tags:
        - Cultivos
      parameters:
        - name: campo
          in: path
          required: true
          description: Tipo de recurso a crear
          schema:
            type: string
            enum: [variedades, modelos, cultivos]
            example: "variedades"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Datos del recurso a crear (estructura variable según el campo)
              $ref: '#/components/schemas/CrearVariedad'
      responses:
        "201":
          description: Recurso creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessCreated'
        "400":
          description: Datos en JSON no proporcionados o campo no soportado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: "Datos de JSON requeridos"
                status: 400
                error: "Bad Request"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example: 
                message: "Provider Error"
                status: 500
                error: "Error del proveedor de datos"
...
components:
  schemas:
...
    CrearVariedad:
      type: object
      properties:
        nombre:
          type: string
          description: Nombre de la variedad
          example: "Vairon"
        cultivo:
          type: string
          description: Nombre del cultivo asociado
          example: "Almendro"
        horas_frio_min:
          type: integer
          example: 200
        horas_frio_max:
          type: integer
          example: 400
        modelo:
          type: string
          description: Codigo publico del modelo
          example: "CM7"
    NivelRiesgoPlaga:
      type: string
      description: Nivel de riesgo de la plaga o enfermedad
      enum: [BAJA, MEDIA, ALTA, MUY_ALTA]
      example: "ALTA"
    TipoAgente:
      type: string
      description: Indica si el agente es una plaga o una enfermedad
      enum: [plaga, enfermedad]
      example: "plaga"
    Plaga:
      type: object
      description: Información de una plaga o enfermedad asociada a un cultivo
      properties:
        nombre:
          type: string
          description: Nombre común de la plaga o enfermedad
          example: "HELMINTOSPORIOSIS"
        agente_causante:
          type: string
          description: Especie o especies responsables de la plaga o enfermedad
          example: "Dreschslera teres"
        momento_critico:
          type: string
          description: Etapa o periodo del cultivo con mayor riesgo de afección
          example: "Desde encañado hasta llenado de grano"
        observaciones:
          type: string
          description: Condiciones ambientales o agronómicas que favorecen su aparición
          example: "Infección en otoño-invierno, explotando en primavera"
        mas_info:
          type: string
          nullable: true
          format: uri
          description: URL con información ampliada sobre la plaga o enfermedad
          example: "https://plagas.itacyl.es/helmintosporisis"
        tipo:
          $ref: '#/components/schemas/TipoAgente'
        nivel_riesgo:
          $ref: '#/components/schemas/NivelRiesgoPlaga'
    CultivoPlagas:
      type: object
      description: Cultivo con su listado de plagas y enfermedades asociadas
      properties:
        cultivo:
          type: object
          description: Información básica del cultivo
          properties:
            nombre:
              type: string
              description: Nombre del cultivo
              example: "Maiz"
            grupo:
              type: string
              description: Grupo o categoría agronómica del cultivo
              example: "cereales"
        plagas:
          type: array
          description: Lista de plagas y enfermedades asociadas al cultivo
          items:
            $ref: '#/components/schemas/Plaga'
    PrediccionPlagasCalculadas:
      type: array
      description: Respuesta con la predicción de plagas calculadas, una entrada por cultivo consultado
      items:
        $ref: '#/components/schemas/CultivoPlagas'
    Modelo:
      type: object
      properties:
        nombre: 
          type: string
          example: "Unidad de Frio (UF)"
        codigo:
          type: string
          example: "UF"
        descripcion:
          type: string
          example: "Acumular horas en las que la temperatura esta comprendida entre 2.5-9.1 grados, si temperatura 1.4-15.9 no suman, temperaturas 16-18 restan"
    VariedadesModelos:
      type: object
      properties:
        nombre_cultivo:
          type: string
          example: "Almendro"
        nombre:
          type: string
          example: "Guara"
        horas_frio_min:
          type: integer
          example: 400
        horas_frio_max:
          type: integer
          example: 600
        modelo:
          $ref: '#/components/schemas/Modelo'  
    VariedadHorasFrio:
      type: object
      properties:
        nombre_variedad:
          type: string
          example: "Lauranne"
        horas_frio_max:
          type: integer
          description: Umbral de horas frio máximas requeridas
          example: 200
        horas_frio_min:
          type: integer
          description: Umbral de horas frio mínimas requeridas
          example: 190
        horas_frio_actuales:
          type: number
          description: Acumulativo actual de las horas de frio sobre la variedad
          example: 682.5
    EtapaFenologica:
      type: object
      properties:
        nombre:
          type: string
          description: Nombre de la etapa etapa fenologica
          example: "dormancia"
        codigo:
          type: string
          description: Codigo publico que identifica a la etapa
          example: "DOR"
        orden:
          type: integer
          description: Indica el orden de la etapa dentro de todas las etapas fenologicas de la variedad
          example: 1
    UmbralesRiesgo:
      type: object
      properties:
        nombre_variedad:
          type: string
          example: "Lauranne"
        etapa_fenologica:
          type: object
          $ref: '#/components/schemas/EtapaFenologica'
        critico:
          type: number
          example: -20.0
        alto:
          type: number
          example: -14.0
        moderado:
          type: number
          example: -5.0
        bajo:
          type: number
          example: 0.0
    SuccessCreated:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Se ha creado correctamente la variedad"
        status:
          type: integer
          example: 201
    SuccessOk:
      type: object
      properties:
        success:
          type: string
          example: "true"
        message:
          type: string
          example: "Variedad Guara actualizada correctamente sus horas de frio"
...
    ApiError:
      type: object
      properties:
        message: 
          type: string
        status:
          type: integer
        error:
          type: string
...